[WIP] Beta version working. The components has been extracted
