+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/components/Footer.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2022. File developed by Jorge Agulló Martín for SoftYorch
+ */
+
+package com.softyorch.taskapp.ui.screensBeta.login.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.softyorch.taskapp.ui.components.ButtonCustom
+
+@Composable
+fun Footer(text: String, enable: Boolean, error: Boolean, onClick: () -> Unit) {
+    Box(
+        modifier = Modifier.fillMaxWidth().padding(top = 16.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        ButtonCustom(
+            text = text,
+            primary = true,
+            enable = enable,
+            error = error
+        ) {
+            onClick()
+        }
+    }
+}
