+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/components/Head.kt
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2022. File developed by Jorge Agulló Martín for SoftYorch
+ */
+
+package com.softyorch.taskapp.ui.screensBeta.login.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.unit.dp
+import com.softyorch.taskapp.utils.extensions.contentColorLabelAsStateAnimation
+
+@Composable
+fun Head(
+    text1: String,
+    text2: String,
+    isClickable: Boolean,
+    onClick: () -> Unit
+) {
+    var click by remember { mutableStateOf(value = false) }
+    val colorText by click.contentColorLabelAsStateAnimation {
+        onClick()
+        click = false
+    }
+    Row(
+        modifier = Modifier
+            .fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.Center
+    ) {
+        Text(
+            modifier = Modifier.padding(top = 4.dp),
+            text = text1,
+            style = MaterialTheme.typography.labelSmall,
+            color = MaterialTheme.colorScheme.onSurface
+        )
+        Text(
+            modifier = Modifier.padding(top = 4.dp)
+                .clickable(enabled = !isClickable) { click = true },
+            text = text2,
+            style = MaterialTheme.typography.labelSmall.copy(
+                textDecoration = TextDecoration.Underline
+            ),
+            color = colorText
+        )
+    }
+}
+
