+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/components/NewAccountBody.kt
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2022. File developed by Jorge Agulló Martín for SoftYorch
+ */
+
+package com.softyorch.taskapp.ui.screensBeta.login.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.softyorch.taskapp.R
+import com.softyorch.taskapp.ui.screensBeta.login.LoginViewModelBeta
+import com.softyorch.taskapp.ui.screensBeta.login.errors.model.ErrorNewAccountModel
+import com.softyorch.taskapp.ui.screensBeta.login.model.NewAccountModel
+
+@Composable
+fun newAccountBody(
+    isBlocked: Boolean,
+    autoLogin: Boolean,
+    viewModel: LoginViewModelBeta,
+    newAccountModel: NewAccountModel,
+    errorsNewAccount: ErrorNewAccountModel,
+    onGo: Boolean
+): Boolean {
+    var onGo1 = onGo
+    Head("¿Ya tienes cuenta? ", "Inicia sesión", (isBlocked || autoLogin)) {
+        viewModel.showNewAccount()
+    }
+    ContentNewAccount(
+        newAccountModel,
+        errorsNewAccount,
+        onGo = { onGo1 = it },
+        viewModel::onNewAccountInputChange
+    )
+    Footer(
+        text = stringResource(R.string.new_account),
+        enable = errorsNewAccount.isActivatedButton,
+        error = errorsNewAccount.error
+    ) {
+        onGo1 = true
+    }
+    return onGo1
+}
